update account balance for each transaction

+++ backend/migrations/versions/bc38ef47ec13_initial_migration.py
@@ -41,7 +41,7 @@ def upgrade():
         sa.Column("id", sa.Integer, primary_key=True),
         sa.Column("account_id", sa.Integer, sa.ForeignKey("account.id"), nullable=False),
         sa.Column("operation", sa.String(20), nullable=False),
-        sa.Column("value", sa.Integer, nullable=False),
+        sa.Column("value", sa.Numeric, nullable=False),
     )
