+++ backend/server/models/transaction.py
@@ -5,7 +5,7 @@ class Transaction(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     account_id = db.Column(db.Integer, db.ForeignKey("account.id"), nullable=False)
     operation = db.Column(db.String(20), nullable=False)
-    value = db.Column(db.Integer, nullable=False)
+    value = db.Column(db.Numeric, nullable=False)
 
     def __repr__(self):
         return f"<Transaction {self.operation} >"
